First functioning version of multi-binding support
It's clunky, and glitchy, but it compiles and runs!

For some reason, the lists seem to be processed in the reverse order
from what I would have expected.  This happened even after I reversed
the order of the list (prepending instead of appending).  Very odd.

Also, OW hotkeys behave poorly, esp. regarding the Any Road stuff.
The handling needs a rethink.